+++ HtmlRun.Common.Runtime/Interfaces/ICurrentInstructionContext.cs
@@ -1,3 +1,4 @@
+using HtmlRun.Common.Models;
 using HtmlRun.Runtime.Code;
 
 namespace HtmlRun.Runtime.Interfaces;