+++ HtmlRun.Common.Runtime/Interfaces/IHeapItem.cs
@@ -1,0 +1,10 @@
+namespace HtmlRun.Runtime.Interfaces;
+
+public interface IHeapItem
+{
+  int Index { get; set; }
+
+  object? Data { get; set; }
+
+  Dictionary<string, object?>? AdditionalData { get; set; }
+}