+++ HtmlRun.Runtime/Factories/CallModelFactory.cs
@@ -1,5 +1,4 @@
 using HtmlRun.Common.Models;
-using HtmlRun.Runtime.Code;
 
 namespace HtmlRun.Runtime.Factories;