+++ HtmlRun.Runtime/HtmlRuntime.cs
@@ -17,9 +17,9 @@ public class HtmlRuntime : IHtmlRuntimeForApp, IHtmlRuntimeForContext, IHtmlRunt
 
   private readonly Dictionary<string, INativeJSDefinition> jsInstructions = new();
 
-  private readonly Stack<Context> ctxStack = new();
-
   private readonly List<PluginBase> plugins = new();
+
+  private readonly Stack<Context> ctxStack;
 
   private readonly Context globalCtx;
 
@@ -29,9 +29,18 @@ public class HtmlRuntime : IHtmlRuntimeForApp, IHtmlRuntimeForContext, IHtmlRunt
 
   private JavascriptParserWithContext? applicationJsContext;
 
-  public HtmlRuntime()
+  public HtmlRuntime(Context? globalCtx = null)
   {
-    this.globalCtx = new Context(null, this.ctxStack);
+    if (globalCtx == null)
+    {
+      this.ctxStack = new Stack<Context>();
+      this.globalCtx = new Context(null, this.ctxStack);
+    }
+    else
+    {
+      this.globalCtx = globalCtx;
+      this.ctxStack = globalCtx.CtxStack;
+    }
   }
 
   public string[] Namespaces => this.instructions.Keys
@@ -179,7 +188,7 @@ public class HtmlRuntime : IHtmlRuntimeForApp, IHtmlRuntimeForContext, IHtmlRunt
 
     //  For information
     string key = $"Plugin.{plugin.Name}";
-    if (!this.globalCtx.IsDeclared(key))
+    if (!this.globalCtx.IsDeclared(key, ignoreInferred: true))
     {
       this.globalCtx.DeclareAndSetConst(key, "true");
     }
@@ -248,12 +257,14 @@ public class HtmlRuntime : IHtmlRuntimeForApp, IHtmlRuntimeForContext, IHtmlRunt
       }
       else if (finalCtx.CursorModification is ContextPull)
       {
-        this.ctxStack.Pop();
+        Context? ctxToDispose = this.ctxStack.Pop();
 
         if (this.ctxStack.Count == 0)
         {
           throw new InvalidOperationException("Context stack is not balanced.");
         }
+
+        ctxToDispose?.Release();
       }
       else
       {
@@ -292,6 +303,11 @@ public class HtmlRuntime : IHtmlRuntimeForApp, IHtmlRuntimeForContext, IHtmlRunt
     if (this.instructions.ContainsKey(key))
     {
       return this.instructions[key];
+    }
+
+    if (this.application == null)
+    {
+      throw new NullReferenceException("Application is null.");
     }
 
     if (this.application.LabeledGroups.Any())
@@ -373,7 +389,7 @@ public class HtmlRuntime : IHtmlRuntimeForApp, IHtmlRuntimeForContext, IHtmlRunt
             ctx.DeclareVariable(resultKey);
           }
 
-          ctx.SetVariable(resultKey, jsResult);
+          ctx.SetValueVariable(resultKey, jsResult);
         };
 
         result[i] = new ParsedArgument(key, ParsedArgumentType.Reference);