+++ HtmlRun.Tests/SQL/NHibernateTests.cs
@@ -129,7 +129,7 @@ public class NHibernateTests : IDisposable
   {
     repo.Insert(session, repo.Create(new Dictionary<string, object?>() { { "Id", 3 }, { "Name", "Test" } }));
 
-    var all = repo.FindAll(session, new Dictionary<string, object>() { { "Id", 3 } }).Cast<dynamic>().ToList();
+    var all = repo.FindAll(session, new Dictionary<string, object?>() { { "Id", 3 } }).Cast<dynamic>().ToList();
     Assert.NotEmpty(all);
     Assert.Equal(3, all[0].Id);
     Assert.Equal("Test", all[0].Name);